+++ src/js/locale/ca-ES.js
@@ -10,6 +10,15 @@
         },
         formatHelpForTagging() {
             return "Per crear les vostres pròpies etiquetes, centreu-vos en el camp d'entrada i feu clic en un dels botons següents per crear l'objecte:";
+        },
+        formatInputToShort(remainingChars) {
+            let message = 'Si us plau, introdueix ' + remainingChars + ' car';
+            if (remainingChars == 1) {
+                message += 'àcter';
+            } else {
+                message += 'àcters';
+            }
+            return message;
         }
     }
     DEFAULTS.locale = "ca-ES";