+++ src/js/locale/es-ES.js
@@ -10,6 +10,15 @@
         },
         formatHelpForTagging() {
             return "Para crear sus propias etiquetas, céntrese en el campo de entrada y haga clic en uno de los siguientes botones para crear el objeto:";
+        },
+        formatInputToShort(remainingChars) {
+            let message = 'Por favor, introduzca ' + remainingChars + ' car';
+            if (remainingChars == 1) {
+                message += 'ácter';
+            } else {
+                message += 'acteres';
+            }
+            return message;
         }
     }
     DEFAULTS.locale = "es-ES";