+++ src/js/locale/et-EE.js
@@ -10,6 +10,14 @@
         },
         formatHelpForTagging() {
             return "Oma siltide loomiseks keskenduge sisestusväljale ja klõpsake objekti loomiseks ühte järgmistest nuppudest:";
+        },
+        formatInputToShort(remainingChars) {
+            let message = 'Sisesta ' + remainingChars + ' täht';
+            if (remainingChars != 1) {
+                message += 'e';
+            }
+            message += ' rohkem';
+            return message;
         }
     }
     DEFAULTS.locale = "et-EE";