+++ src/js/locale/fr-FR.js
@@ -10,6 +10,9 @@
         },
         formatHelpForTagging() {
             return "Pour créer vos propres balises, concentrez-vous sur le champ de saisie et cliquez sur l'un des boutons suivants pour créer l'objet:";
+        },
+        formatInputToShort(remainingChars) {
+            return 'Saisissez au moins ' + remainingChars + ' caractère' + ((remainingChars > 1) ? 's' : '');
         }
     }
     DEFAULTS.locale = "fr-FR";