+++ src/js/locale/ja-JP.js
@@ -10,6 +10,9 @@
         },
         formatHelpForTagging() {
             return "独自のタグを作成するには、入力フィールドに焦点を当て、次のいずれかのボタンをクリックしてオブジェクトを作成します。";
+        },
+        formatInputToShort(remainingChars) {
+            return '少なくとも ' + remainingChars + ' 文字を入力してください';
         }
     }
     DEFAULTS.locale = "ja-JP";