+++ src/js/locale/nl-NL.js
@@ -10,6 +10,9 @@
         },
         formatHelpForTagging() {
             return "Om uw eigen tags te maken, focust u zich op het invoerveld en klikt u op een van de volgende knoppen om het object te maken:";
+        },
+        formatInputToShort(remainingChars) {
+            return 'Gelieve ' + remainingChars + ' of meer karakters in te voeren';
         }
     }
     DEFAULTS.locale = "nl-NL";