+++ src/js/locale/ro-RO.js
@@ -10,6 +10,9 @@
         },
         formatHelpForTagging() {
             return "Pentru a vă crea propriile etichete, concentrați-vă pe câmpul de introducere și faceți clic pe unul dintre următoarele butoane pentru a crea obiectul:";
+        },
+        formatInputToShort(remainingChars) {
+            return 'Vă rugăm să introduceți ' + remainingChars + ' sau mai multe caractere';
         }
     }
     DEFAULTS.locale = "ro-RO";