+++ src/js/locale/zh-CN.js
@@ -10,6 +10,9 @@
         },
         formatHelpForTagging() {
             return "要创建您自己的标签，请关注输入字段并单击以下按钮之一来创建对象：";
+        },
+        formatInputToShort(remainingChars) {
+            return '请再输入至少' + remainingChars + '个字符';
         }
     }
     DEFAULTS.locale = "zh-CN";